backend_estudiantes/requirements.txt
asgiref==3.11.0
cross-web==0.4.1
Django==6.0.1
django-cors-headers==4.9.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
packaging==26.0
promise==2.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
strawberry-graphql==0.289.3
strawberry-graphql-django==0.74.1
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
